Crate vortex_alp

This crate contains an implementation of the floating point compression algorithm from the paper “ALP: Adaptive Lossless floating-Point Compression” by Afroozeh et al.

The compressor has two variants, classic ALP which is well-suited for data that does not use the full precision, and “real doubles”, values that do.

Classic ALP will return small integers, and it is meant to be cascaded with other integer compression techniques such as bit-packing and frame-of-reference encoding. Combined this allows for significant compression on the order of what you can get for integer values.

ALP-RD is generally terminal, and in the ideal case it can represent an f64 is just 49 bits, though generally it is closer to 54 bits per value or ~12.5% compression.

Macros

match_each_alp_float_ptype

Structs

ALP

The array definition trait

ALPArray

ALPEncoding

The array encoding

ALPMetadata

ALPRD

The array definition trait

ALPRDArray

ALPRDEncoding

The array encoding

ALPRDMetadata

Exponents

RDEncoder

Encoder for ALP-RD (“real doubles”) values.

Traits

ALPFloat

ALPRDFloat

Main trait for ALP-RD encodable floating point numbers.

Functions

alp_encode

alp_encode_components

alp_rd_decode

Decode a vector of ALP-RD encoded values back into their original floating point format.

decompress