Struct vortex_alp::ALPRDArray

pub struct ALPRDArray { /* private fields */ }

Implementations

impl ALPRDArray

pub fn len(&self) -> usize

pub fn is_empty(&self) -> bool

impl ALPRDArray

pub fn try_new( dtype: DType, left_parts: Array, left_parts_dict: impl AsRef<[u16]>, right_parts: Array, right_bit_width: u8, left_parts_exceptions: Option<Array>, ) -> VortexResult<Self>

pub fn is_f32(&self) -> bool

Returns true if logical type of the array values is f32.

Returns false if the logical type of the array values is f64.

pub fn left_parts(&self) -> Array

The leftmost (most significant) bits of the floating point values stored in the array.

These are bit-packed and dictionary encoded, and cannot directly be interpreted without the metadata of this array.

pub fn right_parts(&self) -> Array

The rightmost (least significant) bits of the floating point values stored in the array.

pub fn left_parts_exceptions(&self) -> Option<Array>

Patches of left-most bits.

pub fn left_parts_dict(&self) -> &[u16]

The dictionary that maps the codes in left_parts into bit patterns.